+++ aloe/plugin.py
@@ -11,11 +11,8 @@ from builtins import *
 # pylint:enable=redefined-builtin,wildcard-import,unused-wildcard-import
 
 import fnmatch
-import sys
 import os
 import unittest
-
-from nose.plugins.attrib import AttributeSelector
 
 from aloe.fs import FeatureLoader
 from aloe.registry import CALLBACK_REGISTRY
@@ -26,21 +23,15 @@ class GherkinLoader(unittest.loader.TestLoader):
     Collect Gherkin tests.
     """
 
-    # Nose interface has non-Pythonic names
-    # pylint:disable=invalid-name,unused-argument
-
     def __init__(self):
         """Initialise the helper plugin."""
 
-        # Nose has attrib plugin which works as expected but isn't passed the
-        # tests generated from features. Create a local copy which will be used
-        # to veto the tests.
-
         super().__init__()
-        self.attrib_plugin = AttributeSelector()
 
         # Will be set by the main test program
         self.scenario_indices = None
+        self.tags = None
+        self.exclude_tags = None
 
         # Load all the step definitions
         self.feature_dirs = [
@@ -76,35 +67,6 @@ class GherkinLoader(unittest.loader.TestLoader):
             if directory.startswith(feature_dir):
                 return True
 
-    def wantFile(self, file_):
-        """
-        Load features from feature files.
-        """
-
-        # Check that the feature is in one of the features directories
-        file_dir = os.path.abspath(os.path.dirname(file_))
-        if any(
-                file_dir.startswith(os.path.abspath(feature_dir))
-                for feature_dir in self.feature_dirs
-        ):
-            # Check the file extension
-            # Convert to str (not bytes) since Nose passes in both depending on
-            # whether the feature is in a Python module dir or not
-            if isinstance(file_, bytes):
-                file_ = file_.decode(sys.getfilesystemencoding())
-            if os.path.basename(file_).endswith('.feature'):
-                return True
-
-    def wantPython(self, _):
-        """
-        Ignore Python tests if required.
-        """
-
-        if self.ignore_python:
-            return False
-
-    wantClass = wantFunction = wantMethod = wantModule = wantPython
-
     def scenario_matches(self, feature, scenario_index, scenario_name):
         """
         Whether a given scenario is selected by the command-line options.
@@ -118,11 +80,12 @@ class GherkinLoader(unittest.loader.TestLoader):
             if scenario_index not in self.scenario_indices:
                 return False
 
-        if self.attrib_plugin.enabled:
-            scenario = getattr(feature, scenario_name)
-            # False means "no", None means "don't care" for Nose plugins
-            if self.attrib_plugin.validateAttrib(scenario, feature) is False:
-                return False
+        scenario = getattr(feature, scenario_name)
+        tags = getattr(scenario, 'tags', set())
+        if self.tags and not tags & self.tags:
+            return False
+        if self.exclude_tags and tags & self.exclude_tags:
+            return False
 
         return True
 
@@ -136,7 +99,8 @@ class GherkinLoader(unittest.loader.TestLoader):
             for path, _, files in os.walk(file_or_dir):
                 if self.is_feature_directory(path):
                     for filename in fnmatch.filter(files, '*.feature'):
-                        for test in self.tests_from(filename):
+                        filepath = os.path.join(path, filename)
+                        for test in self.tests_from(filepath):
                             yield test
             return
 
@@ -145,17 +109,10 @@ class GherkinLoader(unittest.loader.TestLoader):
         # About to run a feature - ensure "before all" callbacks have run
         self.ensure_before_callbacks()
 
-        has_tests = False
-
         # Filter the scenarios, if asked
         for idx, scenario_name in test.scenarios():
             if self.scenario_matches(test, idx, scenario_name):
-                has_tests = True
                 yield test(scenario_name)
-
-        # Feature OK but no tests filtered
-        if not has_tests:
-            yield False
 
     def ensure_before_callbacks(self):
         """